requirements.txt
pyyaml
pysnmp
